labfiles/foundry-rag/python/requirements.txt
python-dotenv
openai
azure-identity
azure-ai-projects
